Explicit casts to (size_t)
We already handled the case of ssize_t len being -1 or 0, so the cast is safe.

+++ lib/network/tsnetwork_buf.c
@@ -141,13 +141,13 @@ callback_buf(void * cookie, int status)
 		goto docallback;
 	} else {
 		/* Adjust bandwidth limit. */
-		if (network_bwlimit_eat(C->netop, len)) {
+		if (network_bwlimit_eat(C->netop, (size_t)len)) {
 			status = NETWORK_STATUS_ERR;
 			goto docallback;
 		}
 
 		/* Data has been read/written into/from buffer. */
-		C->bufpos += len;
+		C->bufpos += (size_t)len;
 		if (C->bufpos == C->buflen) {
 			status = NETWORK_STATUS_OK;
 			goto docallback;
